[ble] raal 15.4 driver works.

+++ nimble/controller/src/ble_ll_sched.c
@@ -1259,7 +1259,7 @@ ble_ll_sched_next_time(uint32_t *next_event_time)
 uint8_t
 ble_ll_sched_get_current_type(void)
 {
-    if (((int32_t)os_cputime_get32() - (int32_t)g_ble_ll_sched_current.end_time) > 0) {
+    if ((int32_t)(os_cputime_get32() - g_ble_ll_sched_current.end_time) > 0) {
         g_ble_ll_sched_current.type = BLE_LL_SCHED_TYPE_NONE;
     }
